Time: 4 ms (44.77%), Space: 106 MB (7.26%) - LeetHub

+++ 2270-number-of-ways-to-split-array/2270-number-of-ways-to-split-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
 
-        vector<long long> prefix = {nums[0]};
+        vector<long> prefix = {nums[0]};
 
         for(int i = 1; i < nums.size(); i++)
         {
